LayoutTests/platform/mac-wk2/plugins/pdf-plugin-initial-scroll-position-expected.txt

This test ensures that a PDF plugin's initial scroll position is correct (0, 0).

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS internals.pluginScrollPositionX(document.getElementById('embed')) is 0
PASS internals.pluginScrollPositionY(document.getElementById('embed')) is 0
PASS successfullyParsed is true

TEST COMPLETE
